Refactor security rule factories into dedicated module
Moved security rule factories from a general factories file to a focused module `tests/test_factories/security/security_rule.py`. Updated relevant imports and references across the codebase for improved modularity and maintainability.
